Add /exchangeRate endpoint

+++ src/main/java/com/dzvonik/model/Currency.java
@@ -2,12 +2,12 @@ package com.dzvonik.model;
 
 public class Currency {
 
-    private int id;
+    private Long id;
     private String code;
     private String fullName;
     private String sign;
 
-    public Currency(int id, String code, String fullName, String sign) {
+    public Currency(Long id, String code, String fullName, String sign) {
         this.id = id;
         this.code = code;
         this.fullName = fullName;
@@ -20,11 +20,11 @@ public class Currency {
         this.sign = sign;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
